+++ __main__.py
@@ -29,7 +29,7 @@ Modules:
     - scripts: Automation tools
 """
 
-__version__ = "4.3.0"
+__version__ = "4.3.1"
 __author__ = "Deep-Sea Nexus Team"
 
 try: